# docs/openapi.yaml
openapi: 3.0.0
info:
  title: NyaySetu Unified API
  version: 1.0.0
  description: REST API specification for the NyaySetu Digital Judiciary Platform. Covers Spring Boot backend services and Python microservices.
servers:
  - url: http://localhost:8080
    description: Spring Boot Backend Server
  - url: http://localhost:8001
    description: Python Services (NLP Orchestrator / LawGPT)
paths:
...
  /api/v1/client/fir/stats:
    get:
      tags:
        - Client FIRs
      summary: "Get client FIR stats for dashboard"
      description: "Get client FIR stats for dashboard (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
...
  /api/v1/police/fir/list:
    get:
      tags:
        - Police FIRs & Investigations
      summary: "Get all FIRs uploaded by the current officer"
      description: "Get all FIRs uploaded by the current officer (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/fir/pending:
    get:
      tags:
        - Police FIRs & Investigations
      summary: "Get all FIRs pending police review (client-filed FIRs)"
      description: "Get all FIRs pending police review (client-filed FIRs) (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/fir/upload:
    post:
      tags:
        - Police FIRs & Investigations
      summary: "Upload FIR document with SHA-256 digital stamping"
      description: "Upload FIR document with SHA-256 digital stamping (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/fir/{id}:
    get:
      tags:
        - Police FIRs & Investigations
      summary: "Get FIR details by ID"
      description: "Get FIR details by ID (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/fir/{id}/status:
    put:
      tags:
        - Police FIRs & Investigations
      summary: "Update FIR status (REGISTERED or REJECTED)"
      description: "Update FIR status (REGISTERED or REJECTED) (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/fir/{id}/verify:
    post:
      tags:
        - Police FIRs & Investigations
      summary: "Verify FIR integrity by re-hashing uploaded file"
      description: "Verify FIR integrity by re-hashing uploaded file (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
...
  /api/v1/police/investigation/list:
    get:
      tags:
        - Police FIRs & Investigations
      summary: "Get FIRs currently under investigation"
      description: "Get FIRs currently under investigation (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/investigation/{id}/draft-submission:
    get:
      tags:
        - Police FIRs & Investigations
      summary: "Draft Court Submission using Groq (Charge Sheet)"
      description: "Draft Court Submission using Groq (Charge Sheet) (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/investigation/{id}/evidence:
    post:
      tags:
        - Police FIRs & Investigations
      summary: "Upload additional evidence to FIR"
      description: "Upload additional evidence to FIR (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/investigation/{id}/start:
    post:
      tags:
        - Police FIRs & Investigations
      summary: "Start investigation on an FIR"
      description: "Start investigation on an FIR (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/investigation/{id}/submit:
    post:
      tags:
        - Police FIRs & Investigations
      summary: "Submit investigation findings to court"
      description: "Submit investigation findings to court (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/investigation/{id}/summary:
    get:
      tags:
        - Police FIRs & Investigations
      summary: "Generate AI Summary using Groq"
      description: "Generate AI Summary using Groq (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/stats:
    get:
      tags:
        - Police FIRs & Investigations
      summary: "Get police dashboard statistics"
      description: "Get police dashboard statistics (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/summons/pending:
    get:
      tags:
        - Police FIRs & Investigations
      summary: "Get pending summons delivery tasks for police"
      description: "Get pending summons delivery tasks for police (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/police/summons/{caseId}/complete:
    post:
      tags:
        - Police FIRs & Investigations
      summary: "Mark summons as served"
      description: "Mark summons as served (Served by Spring Boot backend)"
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
          description: "The caseId identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Token from /api/v1/auth/login"